feat(*): Improve mobile UI
