+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -4,7 +4,6 @@ about: Create a report to help us improve
 title: ''
 labels: bug
 assignees: ''
-
 ---
 
 ## Describe the bug
@@ -33,6 +32,7 @@ A clear and concise description of what you actual to happen.
 If applicable, add screenshots to help explain your problem.
 
 ## Desktop (please complete the following information)
- - OS: [e.g. iOS]
- - Browser [e.g. chrome, safari]
- - Version [e.g. 22]
+
+- OS: [e.g. iOS]
+- Browser [e.g. chrome, safari]
+- Version [e.g. 22]
