+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -4,7 +4,6 @@ about: Suggest an idea for this project
 title: ''
 labels: ''
 assignees: ''
-
 ---
 
 ## Is your feature request related to a problem? Please describe.
@@ -20,4 +19,5 @@ A clear and concise description of what you want to happen.
 A clear and concise description of any alternative solutions or features you've considered.
 
 ## Additional context
+
 Add any other context or screenshots about the feature request here.
