+++ src/components/Sidebar/internal/Desktop/index.ts
@@ -1,0 +1,1 @@
+export * from './Desktop';
