+++ src/components/Sidebar/internal/Mobile/index.ts
@@ -1,0 +1,1 @@
+export * from './Mobile';
