+++ src/components/Sidebar/internal/SidebarInner/index.ts
@@ -1,0 +1,1 @@
+export * from './SidebarInner';
