+++ src/components/Sidebar/internal/Summary/index.ts
@@ -1,0 +1,1 @@
+export * from './Summary';
