+++ src/components/Sidebar/types.ts
@@ -1,0 +1,4 @@
+export type Props = {
+  inputRef?: React.RefObject<HTMLInputElement>;
+  listRef?: React.RefObject<HTMLDivElement>;
+};
