+++ src/components/Viewer/internal/ComparisonView/TwoUp.tsx
@@ -1,14 +1,18 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Image } from '../../../Image';
-import { Space } from '../../../../styles/variables';
+import { Space, BreakPoint } from '../../../../styles/variables';
 
 const Wrapper = styled.div`
   display: grid;
-  grid-template-columns: 50% 50%;
-  grid-gap: ${Space * 3}px;
+  grid-template-columns: repeat(2, 1fr);
+  grid-gap: ${Space * 1}px;
   justify-content: center;
   align-items: center;
+
+  @media (min-width: ${BreakPoint.MEDIUM}px) {
+    grid-gap: ${Space * 3}px;
+  }
 `;
 
 export type Props = {
